TypeScript for utilities
