+++ package.json
@@ -35,6 +35,12 @@
     },
     "devDependencies": {
         "@rdfjs/types": "^1.0.1",
-        "mocha": "^8.3.2"
+        "@types/mocha": "^9.0.0",
+        "@types/n3": "^1.10.3",
+        "@types/rdfjs__namespace": "^1.1.5",
+        "mocha": "^8.3.2",
+        "ts-mocha": "^8.0.0",
+        "ts-node": "^10.2.1",
+        "typescript": "^4.3.5"
     }
 }
