+++ src/prec/rules-for-properties.js
@@ -4,7 +4,7 @@ const N3 = require('n3');
 const namespace = require('@rdfjs/namespace');
 
 const DStar    = require('../dataset/index.js');
-const QuadStar = require('../rdf/quad-star.js');
+const QuadStar = require('../rdf/quad-star');
 
 const rdf  = namespace("http://www.w3.org/1999/02/22-rdf-syntax-ns#", N3.DataFactory);
 const rdfs = namespace("http://www.w3.org/2000/01/rdf-schema#"      , N3.DataFactory);
